banned

+++ bot.py
@@ -42,7 +42,7 @@ ImMEE7 = ['MEE7, not MEE6, it\'s MEE7. Don\'t you dare mix us up',
           'MEE7? That\'s me baby!! Don\'t wear it out 😉']
 
 Acceptance_Emojis = ['😍', '❤️', '🥰', '💪', '👑', '☺️', '🤙']
-banned = ['vermont', 'green mountain state', 'v e r m o n t']
+banned = ['vermont', 'green mountain state', 'v e r m o n t','💌','❣️','💓','💕','💗','💘','💙','💚','💖','💛','💜','💝','💞','💟','🧡','🏩','👩‍❤️‍👨','❤️','👩‍❤️‍💋‍👨','🖤','♥️','😍','🤍','🤎','😘','😻','🥰']
 
 
 def gen_ID(char):
@@ -179,14 +179,14 @@ async def on_message(message):
     # banned  = firebase.get('/' + FIREBASE_NAME + '/banned/'+ str(message.guild), '')
     CENSOR_DICT = censorship()
     Server = str(message.guild.id)
-
-    for i in banned:
-        if i in message.content.lower():
-            await message.delete()
-            if CENSOR_DICT[Server]:
-                await message.channel.send(f'*{i}* is banned please shut the f@@k up already {message.author.mention}')
-            else:
-                await message.channel.send(f'*{i}* is banned please shut the fuck up already {message.author.mention}')
+    if message.guild.id == '684944796779151406':
+        for i in banned:
+            if i in message.content.lower():
+                await message.delete()
+                if CENSOR_DICT[Server]:
+                    await message.channel.send(f'*{i}* is banned please shut the f@@k up already {message.author.mention}')
+                else:
+                    await message.channel.send(f'*{i}* is banned please shut the fuck up already {message.author.mention}')
 
     if message.author == MEE6:
         Server = str(message.guild.id)
